WordCamp Organizer Reminders: Change definition of 'added to schedule' trigger.
Previously it was defined as when a published `wordcamp` post added their start date. Now we're defining it as when
a camp moves from the 'pending' status to 'publish'. This makes it more consistent with current practices (e.g.,
the start date and other meta fields are required before a camp is allowed to go from pending to publish.)


git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins/wordcamp-organizer-reminders@666 74240141-8908-4e6f-9713-ba540dce6ec7